feat: WIP using dataclasses
fix: updating test case files

fix: most tests passing

fix: removed unnecessary functions

fix: tests passing

fix: moved test case files
