+++ dpypelines/pipeline/errors/__init__.py
@@ -1,13 +1,11 @@
 from .dataset_api_request_creation_exception import DatasetAPIRequestCreationException
 from .dataset_not_found_exception import DatasetNotFoundException
 from .dataset_type_exception import DatasetTypeException
-from .distributions_exception import DistributionsException
 from .validation_exception import ValidationException
 
 all = [
     DatasetAPIRequestCreationException,
     DatasetNotFoundException,
     DatasetTypeException,
-    DistributionsException,
     ValidationException,
 ]
