+++ dpypelines/pipeline/validate_pipeline.py
@@ -1,74 +1,80 @@
 import json
-import re
 from pathlib import Path
-from typing import Dict, List
 
 from dpytools.logging.logger import DpLogger
+from dpytools.stores.directory.local import LocalDirectoryStore
 from dpytools.validation.json.validation import validate_json_schema
 
-from dpypelines.pipeline.configuration import get_pipeline_config
-from dpypelines.pipeline.messages.utils import get_submitter_email
-from dpypelines.pipeline.pipelineconfig.matching import get_matching_pattern
+from dpypelines.pipeline.models import Manifest, Metadata
 
 logger = DpLogger("data-ingress-pipeline")
 
 
-def validate_pipeline_files(files_dir: Path, pipeline_config: dict) -> Dict:
+def retrieve_and_validate_manifest(
+    local_store: LocalDirectoryStore,
+) -> Manifest:
+    """Retrieve and validate manifest from the local directory store."""
+    manifest_dict = local_store.get_lone_matching_json_as_dict("manifest.json")
+
+    if not manifest_dict:
+        err_msg = "Failed to retrieve manifest from the local directory store."
+        raise FileNotFoundError(err_msg)
+
+    # Validate manifest against JSON schema
+    validate_manifest_schema(manifest_dict)
+
+    logger.info(
+        "Manifest retrieved and validated against schema.",
+        data={
+            "manifest_dict": manifest_dict,
+        },
+    )
+
+    # Create Manifest object
+    manifest = Manifest.model_validate(manifest_dict)
+    return manifest
+
+
+def validate_manifest_schema(manifest_dict: dict) -> None:
+    """Validate manifest dictionary against the schema."""
+    try:
+        file_path = Path(__file__).parent.parent
+        schema_path = file_path / "schemas" / "manifest_v1_schema.json"
+        validate_json_schema(
+            schema_path=schema_path,
+            data_dict=manifest_dict,
+            error_msg="Invalid manifest",
+        )
+    except Exception as e:
+        raise ValueError(f"Manifest schema validation failed: {e}")
+
+
+def validate_pipeline_files(
+    manifest: Manifest, local_store: LocalDirectoryStore
+) -> Metadata:
     """
     Main validation function that returns validated objects.
     """
-    required_keys = ["manifestVersion", "source_id", "fileAuthorEmail"]
-    # 1. Check core required files
-    required_files = ["metadata.json", "manifest.json"]
-    for file_name in required_files:
-        validate_file_exists_and_not_empty(files_dir / file_name)
-
-    # 2. Retrieve and validate manifest.json and metadata.json
-    manifest_dict = retrieve_and_validate_manifest(
-        files_dir / "manifest.json", required_keys
+    # Retrieve and validate metadata.json
+    validate_file_exists_and_not_empty(
+        local_store.local_path.absolute() / manifest.metadata_file
     )
-    metadata_dict = validate_json_file(files_dir / "metadata.json")
-
-    # 3. Validate config-required files
-    config_files = []
-    config_files.extend(
-        validate_pattern_files(files_dir, pipeline_config, "required_files")
+    metadata_dict = validate_json_file(
+        local_store.local_path.absolute() / manifest.metadata_file
     )
 
-    # 4. Validate supplementary files
-    supplementary_files = validate_pattern_files(
-        files_dir, pipeline_config, "supplementary_distributions"
-    )
-    config_files.extend(supplementary_files)
+    # Create Metadata model from metadata_dict
+    metadata = Metadata.model_validate(metadata_dict)
 
-    return {
-        "manifest": manifest_dict,
-        "metadata": metadata_dict,
-        "input_files": config_files,
-        "config_files": config_files,
-        "supplementary_files": supplementary_files,
-    }
+    # Retrieve and validate data files
+    for distribution in metadata.distributions:
+        validate_file_exists_and_not_empty(
+            local_store.local_path.absolute() / distribution.file
+        )
+        logger.info("Data file found", data={"data_file_name": distribution.file})
 
-
-def validate_pattern_files(
-    files_dir: Path, pipeline_config: dict, pattern_key: str
-) -> List[Path]:
-    """Validate files matching a regex pattern exist and are not empty."""
-    collected_files = []
-    patterns = get_matching_pattern(pipeline_config, pattern_key)
-
-    if patterns:
-        for pattern in patterns:
-            regex = re.compile(pattern)
-            matched_files = [f for f in files_dir.iterdir() if regex.match(f.name)]
-            if not matched_files:
-                raise FileNotFoundError(f"No files found matching pattern: {pattern}")
-
-            for file in matched_files:
-                validate_file_exists_and_not_empty(file)
-                collected_files.append(file)
-
-    return collected_files
+    logger.info("Metadata and data files validated.")
+    return metadata
 
 
 def validate_file_exists_and_not_empty(file_path: Path) -> None:
@@ -87,98 +93,3 @@ def validate_json_file(file_path: Path) -> dict:
         return data
     except json.JSONDecodeError as e:
         raise ValueError(f"File is not valid JSON: {str(e)}")
-
-
-def validate_manifest_vars(manifest_dict: dict, required_keys: list) -> None:
-    """Validate manifest dictionary has required fields."""
-    missing_keys = [key for key in required_keys if key not in manifest_dict]
-
-    if missing_keys:
-        raise KeyError(f"Missing required keys in manifest: {', '.join(missing_keys)}")
-
-    # Validate submitter email
-    get_submitter_email(manifest_dict)
-
-
-def validate_manifest_schema(manifest_dict: dict) -> None:
-    """Validate manifest dictionary against the schema."""
-    try:
-        file_path = Path(__file__).parent.parent
-        schema_path = file_path / "schemas" / "manifest_v1_schema.json"
-        validate_json_schema(
-            schema_path=schema_path,
-            data_dict=manifest_dict,
-            error_msg="Invalid manifest",
-        )
-    except Exception as e:
-        raise ValueError(f"Manifest schema validation failed: {e}")
-
-
-def retrieve_and_validate_manifest(manifest_path: Path, required_keys: list) -> dict:
-    """Retrieve and validate the manifest.json file."""
-    try:
-        manifest_dict = validate_json_file(manifest_path)
-        validate_manifest_vars(manifest_dict, required_keys)
-        validate_manifest_schema(manifest_dict)
-        return manifest_dict
-    except Exception as e:
-        raise ValueError(f"Failed to retrieve and validate manifest: {str(e)}")
-
-
-def retrieve_config_and_files(local_store):
-    """Retrieve configuration and files from the local directory."""
-    manifest_dict = retrieve_manifest(local_store)
-    source_id = get_source_id_from_manifest(manifest_dict)
-    pipeline_config = get_pipeline_config_for_source(source_id)
-    files_dir = Path(local_store.get_current_source_pathlike())
-
-    if not manifest_dict or not source_id or not pipeline_config or not files_dir:
-        err_msg = "Failed to retrieve configuration and files from the local directory."
-        raise FileNotFoundError(err_msg)
-
-    logger.info(
-        "Configuration and files retrieved successfully",
-        data={
-            "manifest_dict": manifest_dict,
-            "source_id": source_id,
-            "pipeline_config": pipeline_config,
-            "files_dir": str(files_dir),
-        },
-    )
-
-    return manifest_dict, pipeline_config, files_dir
-
-
-def retrieve_manifest(local_store):
-    """Retrieve the manifest.json file from the local directory."""
-    manifest_dict = local_store.get_lone_matching_json_as_dict("manifest.json")
-
-    if not manifest_dict:
-        err_msg = "manifest.json not found in the local store."
-        raise FileNotFoundError(err_msg)
-    return manifest_dict
-
-
-def get_source_id(manifest_dict: dict) -> str:
-    """
-    This function returns the `source_id` form the provided manifest_dict (which is the data in the manifest.json file).
-    """
-    return manifest_dict["source_id"]
-
-
-def get_source_id_from_manifest(manifest_dict):
-    """Extract the source_id from the manifest."""
-    source_id = get_source_id(manifest_dict)
-    if not source_id:
-        err_msg = f"source_id not found within manifest: {manifest_dict}."
-        raise KeyError(err_msg)
-    return source_id
-
-
-def get_pipeline_config_for_source(source_id):
-    """Retrieve pipeline configuration using source_id."""
-    pipeline_config = get_pipeline_config(source_id)
-    if not pipeline_config:
-        err_msg = f"Pipeline configuration not found for source_id: {source_id}."
-        raise ValueError(err_msg)
-    return pipeline_config
