+++ pyproject.toml
@@ -13,6 +13,7 @@ unidecode = "^1.3.8"
 xmltodict = "^0.14.2"
 python-dotenv = "^1.1.0"
 dpytools = { git = "https://github.com/ONSdigital/dp-python-tools.git", tag = "v0.7.0" }
+pydantic = "^2.11.3"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.5"
