--- tests/fixtures/test-cases/dataset_ingress_v1/invalid_no_metadata/manifest.json
@@ -1,11 +1,0 @@
-{
-    "manifestVersion": 1,
-    "source_id": "test_move",
-    "fileAuthorEmail": "test@email.com",
-    "fileAuthorUsername": "Username",
-    "isPublishable": false,
-    "licence": "My licence",
-    "licenceUrl": "http://www.example.org/licence",
-    "title": "4g Coverage",
-    "aliasName": "4g-coverage"
-}
