+++ tests/fixtures/test-cases/invalid_no_data/manifest.json
