+++ tests/fixtures/test-cases/invalid_no_data/metadata.json
