+++ tests/fixtures/test-cases/invalid_no_manifest/metadata.json
