+++ tests/fixtures/test-cases/invalid_no_metadata/manifest.json
