+++ tests/fixtures/test-cases/invalid_no_supplementary/manifest.json
