+++ tests/fixtures/test-cases/invalid_no_supplementary/metadata.json
