+++ tests/fixtures/test-cases/test_metadata_dataset_id_missing.json
@@ -1,5 +1,4 @@
 {
-    "dataset_id": "test-static-dataset-1",
     "edition": "test-edition-1",
     "edition_title": "July to September 2022",
     "quality_designation": "accredited-official",
