--- tests/fixtures/test-cases/test_metadata_invalid.json
@@ -1,61 +1,0 @@
-{
-    "@context": {
-        "adms": "http://www.w3.org/ns/adms#",
-        "dcat": "http://www.w3.org/ns/dcat#",
-        "dcterms": "http://purl.org/dc/terms/",
-        "dqv": "http://www.w3.org/ns/dqv#",
-        "foaf": "http://xmlns.com/foaf/0.1/",
-        "vcard": "http://www.w3.org/2006/vcard/ns#"
-    },
-    "@type": "dcat:DatasetSeries",
-    "dcterms:title": "Dataset title",
-    "dcterms:description": "Dataset description",
-    "TBC:nextRelease": "The next release date for a dataset",
-    "dcat:theme": [
-        "theme1",
-        "theme2"
-    ],
-    "dqv:hasQualityAnnotation": [
-        {
-            "dqv:QualityCertificate": "qmi.href.string_required"
-        }
-    ],
-    "dcat:contactPoint": [
-        {
-            "vcard:fn": "contact.name.string_required",
-            "vcard:hasEmail": "contact.email.string_required",
-            "vcard:hasTelephone": "contact.telephone.string"
-        }
-    ],
-    "dcterms:publisher": [
-        {
-            "TBC:type": "publisher.type.string_required",
-            "foaf:name": "publisher.name.string_required",
-            "foaf:homepage": "publisher.href.string_required"
-        }
-    ],
-    "TBC:edition": {
-        "@type": "dcat:Dataset",
-        "dcterms:identifier": "time-series",
-        "TBC:quality_designation": "QualityDesignationEnum",
-        "TBC:usage_notes": [
-            {
-                "TBC:title": "usage_notes.title.string",
-                "TBC:note": "usage_notes.note.string"
-            }
-        ],
-        "TBC:alerts": [
-            {
-                "TBC:type": "AlertTypeEnum",
-                "adms:versionNotes": "alerts.description.string_required"
-            }
-        ],
-        "dcat:distribution": [
-            {
-                "dcterms:title": "distribution.title.string_required",
-                "dcat:mediaType": "MediaTypeEnum",
-                "TBC:distributionFormat": "DistributionFormatEnum"
-            }
-        ]
-    }
-}
