+++ tests/fixtures/test-cases/valid/manifest.json
