+++ tests/fixtures/test-cases/valid/metadata.json
